requirements.txt
django==2.2.3
uuid==1.30
djangorestframework==3.9.4
